resource/requirements.txt
fastapi
gunicorn
asyncpg
pydantic-sqlalchemy
python-dotenv
requests
passlib[bcrypt]
python-jose[cryptography]
PyJWT
vakt
python-multipart
fastapi-cache2[redis]
celery
flower
redis
alembic
uvicorn[standard]